backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
requests
duckduckgo_search
psycopg2-binary
bcrypt
textblob
pyjwt
praw
